chatbot-server/requirements.txt
fastapi[all]
uvicorn
pydantic
httpx
requests
python-dotenv
groq
langchain-chroma
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
kafka-python==1.4.7
kafka
aiokafka
bs4
langgraph
langgraph-checkpoint-sqlite